server/requirements.txt
pipecat-ai[silero,ollama,websocket,whisper]
torch
numpy
loguru
